fix lvl2 whitespace tok
before this multiple mergable tokens in a row,
meaning multiple tokens that are not seperated by
whitespaces, would fail to merge into one. reason:
after the first merge, the next token will be of
toktype TOK_NULL. So when that token becomes
current the next won't be merged to the prev, as
it should be. but that's fixed now.